move BatchableMixin to common.py
